backend/requirements.txt
flask
flask-cors
openai
python-dotenv
firebase-admin
typing_extensions
google.cloud
google-cloud-firestore
google-cloud-storage
google-cloud-secret-manager
google-cloud-core
stripe